+++ demo/pages/DatePicker/index.js
@@ -23,4 +23,11 @@ Page({
   handleChangeDate() {
     this.setData({ defaultDate: new Date('2019/05/02') });
   },
+  handleTriggerPicker(visible, e) {
+    console.log('onTriggerPicker', visible, e);
+  },
+
+  handleDismiss(e) {
+    console.log('e', e)
+  }
 });