+++ src/DatePicker/props.d.ts
@@ -84,5 +84,10 @@ export interface IDatePickerProps extends IBaseFormItemPropsWithOutFocus {
    * @description 切换显示隐藏
    */
   onTriggerPicker?: (visible, e: Record<string, any>) => void;
+  /**
+   * @description 点击蒙层是否可以关闭
+   * @default false
+   */
+  maskClosable?: boolean;
 }
 export declare const DatePickerDefaultProps: Partial<IDatePickerProps>;