+++ src/Rate/props.js
@@ -1,0 +1,10 @@
+export const RateDefaultProps = {
+    maxRate: 5,
+    initRate: 0,
+    allowHalf: false,
+    readOnly: false,
+    disabled: false,
+    halfInactiveColor: "#DBDBDB",
+    activeColor: '#FFD24A',
+    inactiveColor: '#E5E5E5'
+};